# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for a user API.",
        "title": "User API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/patient/register": {
            "post": {
                "description": "Register a new patient in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "patients"
                ],
                "summary": "Register a new patient",
                "parameters": [
                    {
                        "description": "Patient registration data",
                        "name": "patient",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PatientRegisterPatientRequestDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Patient registered successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.PatientRegisterResponseDto"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to register user",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.PatientRegisterPatientRequestDto": {
            "type": "object",
            "required": [
                "first_name",
                "gender",
                "hospital_id",
                "last_name",
                "password"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "allergies": {
                    "type": "string"
                },
                "birth_date": {
                    "type": "string"
                },
                "blood_type": {
                    "type": "string"
                },
                "emergency_contact": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string",
                    "enum": [
                        "male",
                        "female",
                        "other"
                    ]
                },
                "hospital_id": {
                    "description": "Patient specific fields",
                    "type": "string"
                },
                "id_card_number": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
        "dto.PatientRegisterResponseDto": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    }
}